backend/requirements.txt
# Django Core
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.0
django-filter==23.5

# Authentication
djangorestframework-simplejwt==5.3.0
setuptools>=65.0.0

# Database
psycopg2-binary==2.9.9
dj-database-url==2.1.0

# Cloud Services
google-cloud-storage==2.10.0
google-cloud-pubsub==2.18.4
google-cloud-logging==3.8.0

# Redis
redis==5.0.1
django-redis==5.4.0

# Environment
python-dotenv==1.0.0

# API Documentation
drf-spectacular==0.26.5

# Utilities
Pillow==10.1.0
python-dateutil==2.8.2
reportlab==4.0.7

# WSGI Server
gunicorn==21.2.0

# Machine Learning
scikit-learn==1.7.2
joblib==1.5.2
pandas==2.3.3
numpy==2.3.4

# Development & Testing
ipython==8.17.2
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
factory-boy==3.3.0

# Logging
python-json-logger==2.0.7
